// packages/source-sync/src/lambda-modules/sync-boundary/infrastructure/adapter/out/db/typeorm/fields/bt-document-columns.enum.ts
export enum BtDocumentColumns {
  ID = 'ID',
  NEXUS_DOCUMENT_NUMBER = 'NEXUS_DOCUMENT_NUMBER',
  CUT_START_AT = 'CUT_START_AT',
  WEATHER_CONDITION = 'WEATHER_CONDITION',
  REGION = 'REGION',
  OPERATIVE_AREA = 'OPERATIVE_AREA',
  COUNTY = 'COUNTY',
  LOCALITY = 'LOCALITY',
  INITIAL_AFFECTED_CUSTOMERS = 'INITIAL_AFFECTED_CUSTOMERS',
  AFFECTED_CUSTOMERS = 'AFFECTED_CUSTOMERS',
  TOTAL_CLAIMS_COUNT = 'TOTAL_CLAIMS_COUNT',
  DELETED_AT = 'DELETED_AT',
  CREATED_AT = 'CREATED_AT',
  UPDATED_AT = 'UPDATED_AT',
}